feat: filter works by status

+++ packages/api/src/modules/people-works/resolvers.js
@@ -1,12 +1,14 @@
 import models from '../../setup/models';
 
 export async function insertStaff(workId, staff) {
-  await models.PeopleWorks.destroy({ where: { workId: workId } });
+  if (staff) {
+    await models.PeopleWorks.destroy({ where: { workId: workId } });
 
-  const staffs = staff.map(staff => ({
-    workId,
-    peopleId: staff.people.id,
-    rol: staff.rol
-  }));
-  return await models.PeopleWorks.bulkCreate(staffs, { returning: true });
+    const staffs = staff.map(staff => ({
+      workId,
+      peopleId: staff.people.id,
+      rol: staff.rol
+    }));
+    return await models.PeopleWorks.bulkCreate(staffs, { returning: true });
+  }
 }
